+++ numpy/__init__.pyi
@@ -329,6 +329,7 @@ from numpy._core.multiarray import (
     fromstring,
     frompyfunc,
     flatiter,
+    nditer,
     nested_iters,
     flagsobj,
 )
@@ -6007,96 +6008,6 @@ class broadcast:
     def __next__(self) -> tuple[Any, ...]: ...
     def __iter__(self) -> Self: ...
     def reset(self) -> None: ...
-
-@final
-class nditer:
-    @overload
-    def __init__(
-        self,
-        /,
-        op: ArrayLike,
-        flags: Sequence[_NDIterFlagsKind] | None = None,
-        op_flags: Sequence[_NDIterFlagsOp] | None = None,
-        op_dtypes: DTypeLike | None = None,
-        order: _OrderKACF = "K",
-        casting: _CastingKind = "safe",
-        op_axes: Sequence[SupportsIndex] | None = None,
-        itershape: _ShapeLike | None = None,
-        buffersize: SupportsIndex = 0,
-    ) -> None: ...
-    @overload
-    def __init__(
-        self,
-        /,
-        op: Sequence[ArrayLike | None],
-        flags: Sequence[_NDIterFlagsKind] | None = None,
-        op_flags: Sequence[Sequence[_NDIterFlagsOp]] | None = None,
-        op_dtypes: Sequence[DTypeLike | None] | None = None,
-        order: _OrderKACF = "K",
-        casting: _CastingKind = "safe",
-        op_axes: Sequence[Sequence[SupportsIndex]] | None = None,
-        itershape: _ShapeLike | None = None,
-        buffersize: SupportsIndex = 0,
-    ) -> None: ...
-
-    def __enter__(self) -> nditer: ...
-    def __exit__(
-        self,
-        exc_type: type[BaseException] | None,
-        exc_value: BaseException | None,
-        traceback: TracebackType | None,
-    ) -> None: ...
-    def __iter__(self) -> nditer: ...
-    def __next__(self) -> tuple[NDArray[Any], ...]: ...
-    def __len__(self) -> int: ...
-    def __copy__(self) -> nditer: ...
-    @overload
-    def __getitem__(self, index: SupportsIndex) -> NDArray[Any]: ...
-    @overload
-    def __getitem__(self, index: slice) -> tuple[NDArray[Any], ...]: ...
-    def __setitem__(self, index: slice | SupportsIndex, value: ArrayLike) -> None: ...
-    def close(self) -> None: ...
-    def copy(self) -> nditer: ...
-    def debug_print(self) -> None: ...
-    def enable_external_loop(self) -> None: ...
-    def iternext(self) -> py_bool: ...
-    def remove_axis(self, i: SupportsIndex, /) -> None: ...
-    def remove_multi_index(self) -> None: ...
-    def reset(self) -> None: ...
-    @property
-    def dtypes(self) -> tuple[dtype, ...]: ...
-    @property
-    def finished(self) -> py_bool: ...
-    @property
-    def has_delayed_bufalloc(self) -> py_bool: ...
-    @property
-    def has_index(self) -> py_bool: ...
-    @property
-    def has_multi_index(self) -> py_bool: ...
-    @property
-    def index(self) -> int: ...
-    @property
-    def iterationneedsapi(self) -> py_bool: ...
-    @property
-    def iterindex(self) -> int: ...
-    @property
-    def iterrange(self) -> tuple[int, ...]: ...
-    @property
-    def itersize(self) -> int: ...
-    @property
-    def itviews(self) -> tuple[NDArray[Any], ...]: ...
-    @property
-    def multi_index(self) -> tuple[int, ...]: ...
-    @property
-    def ndim(self) -> int: ...
-    @property
-    def nop(self) -> int: ...
-    @property
-    def operands(self) -> tuple[NDArray[Any], ...]: ...
-    @property
-    def shape(self) -> tuple[int, ...]: ...
-    @property
-    def value(self) -> tuple[NDArray[Any], ...]: ...
 
 def from_dlpack(
     x: _SupportsDLPack[None],
